feat:移动进阶查询需求到advancedUserController里，添加一部分按钮及其报错在userSuccess导览界面

+++ src/main/java/com/example/pj/mapper/FavoriteMapper.java
@@ -59,5 +59,5 @@ public interface FavoriteMapper {
             "AND o.OrderDate >= DATE_SUB(NOW(), INTERVAL #{timePeriod}) " +
             "GROUP BY d.DishID, d.DishName " +
             "ORDER BY d.DishID")
-    List<DishSales> getFavoriteDishSales(@Param("userId") int userId, @Param("timePeriod") String timePeriod);
+    List<DishSales> getFavoriteDishSales(@Param("userId") Long userId, @Param("timePeriod") String timePeriod);
 }
